Make sure v.rm returns a promise since it can be async at the moment, log out the output from a v.exec() call. Snapshot update.

+++ volo/volo/v.js
@@ -46,14 +46,21 @@ define(function (require) {
                                             (encoding || defaultEncoding));
                 },
                 rm: function (dirOrFile) {
+                    var d, stat;
+
                     dirOrFile = resolve(dirOrFile);
-                    var stat = fs.statSync(dirOrFile);
-                    if (stat.isFile()) {
-                        fs.unlinkSync(dirOrFile);
-                    } else if (stat.isDirectory()) {
-                        //TODO: need to make rmdir synchronous
-                        file.rmdir(dirOrFile);
+                    if (dirOrFile) {
+                        stat = fs.statSync(dirOrFile);
+                        if (stat.isFile()) {
+                            fs.unlinkSync(dirOrFile);
+                        } else if (stat.isDirectory()) {
+                            //TODO: need to make rmdir synchronous
+                            return file.rmdir(dirOrFile);
+                        }
                     }
+                    d = q.defer();
+                    d.resolve();
+                    return d.promise;
                 },
                 mv: function (start, end) {
                     return fs.renameSync(start, end);
@@ -106,7 +113,7 @@ define(function (require) {
                             if (error) {
                                 d.reject(error);
                             } else {
-                                d.resolve();
+                                d.resolve(stdout);
                             }
                         }
                     );
